chore(*): refactor storage into separate project
