+++ Micro.Starter.Api/Program.cs
@@ -4,8 +4,8 @@ using App.Metrics;
 using App.Metrics.AspNetCore;
 using App.Metrics.Extensions.Configuration;
 using App.Metrics.Formatters.InfluxDB;
-using Micro.Starter.Api.Models;
-using Micro.Starter.Api.StartupExtensions;
+using Micro.Starter.Api.Internal.StartupExtensions;
+using Micro.Starter.Storage;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.Extensions.Configuration;
