+++ Micro.Starter.Common/Uuid/IUuidService.cs
@@ -1,4 +1,4 @@
-namespace Micro.Starter.Api.Uuid
+namespace Micro.Starter.Common.Uuid
 {
     public interface IUuidService
     {
