+++ Micro.Starter.Storage/WeatherRepository.cs
@@ -1,12 +1,20 @@
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
-using Micro.Starter.Api.Models;
-using Micro.Starter.Api.Uuid;
+using Micro.Starter.Common.Uuid;
 using Microsoft.EntityFrameworkCore;
 
-namespace Micro.Starter.Api.Repository
+namespace Micro.Starter.Storage
 {
+    public interface IWeatherRepository
+    {
+        Task<IEnumerable<Weather>> GetAll();
+        Task<Weather> FindById(string id);
+        Task<Weather> Create([NotNull] Weather weather);
+        Task Delete(string id);
+    }
+
     public class WeatherRepository : IWeatherRepository
     {
         private readonly ApplicationContext _db;
